api: do not log controller-runtime messages after context cancel (#5213)
This is mostly annoying in tests when trying to reproduce an issue
and running with something like `-count 100`. With the async nature
of the reconcilers, it's pretty inevitable that some things will
still be in-flight when the tests are done (and so cancel the context
to clean up resources), and a bunch of irrelevant errors that are
really just wrapped "context canceled" get logged out.