kubernetesapply: improve cmd timeout handling
If the process is killed due to context timeout, it'll have a
non-zero exit code (typically 137).

In these instances, we'll let the user know the command took too
long and point them at the Tiltfile API for `update_settings` so
they can increase the timeout value if needed.

Additionally, if stdout is empty on failure (which, given that
the commands are only supposed to write YAML on success to it, it
generally will be), it won't be output, so you don't get a blank
`stdout:` in the logs.