tiltfile: ignore readiness probe for local resource w/ no serve_cmd
If a local resource has no `serve_cmd`, it cannot logically have a
readiness probe, as there's no server to check for readiness.

The actual impact of this was that if another resource had a
dependency on a local resource in this state, the dependent resource
would get stuck waiting on the local resource indefinitely.

Now, the readiness probe will be ignored and a warning logged.