tiltfile: infer object properties during load
As we move to the apiserver being the source of truth, a greater
degree of objects and their fields are managed by the reconciler
based on watches for other objects.

To facilitate this transition, there are several `InferXXX()`
helpers on `Manifest`, which are called during reconciliation for
a `Tiltfile`.

However, this meant that the objects were an inconsistent (at
best) state before reconciliation stabilized. This caused issues
in some cases, the most notable being that `tilt docker-prune`
had stopped working because image name inference hadn't occurred
(it does not rely on a full apiserver + reconciliation).

Now, Live Update + image properties are both inferring during the
Tiltfile load (based on the values in the Tiltfile) in addition to
the reconciler, which resolves issues with Docker pruning.