web: fix resource filtering infinite loop (#5249)
This is a temporary change to store the resource list options in
session storage instead of local storage. This should prevent odd
behavior when multiple tabs are open to the same Tilt instance and
filtering is done in one which causes an infinite loop where Tilt
phantom types characters into the filter.

Longer-term, because we use local storage elsewhere, we need to
investigate the issues with the React library and/or our usage of
it to avoid these types of problems.

For filtering, having it be synced across tabs/persistent across
Tilt sessions is not critical, so using session storage is a
reasonable mitigation for now.