k8s: wait for delete to be complete
This adapts the code originally added in #4626 to apply to all
delete cases. Previously, it was only done in `deleteAndCreate`.
However, when `Delete` is called explicitly, it's done as part of
a resource trigger/full rebuild, so the same eventual issue applies
where Tilt is going to (almost) immediately re-create the entities
it just deleted and might hit conflicts as a result.