controllers: add cluster client cache (#5284)
The `ConnectionManager` is the internal struct used by the
Cluster reconciler and has a `ConnectionCache` interface for use
by the other reconcilers that allows read-only access to the
cached Kubernetes/Docker client for a Cluster object.

Once wired up, dependent reconcilers will still be expected to set
up a watch on the Cluster and re-evaluate when it changes to handle
runtime changes to the Cluster config. However, they can all share
the same underlying client(s), as they do now, to reduce number of
connections and share any client-level caching.