cluster: refactor caching and add configuration hash
This is an attempt at a middle-ground that doesn't totally sidestep
the apiserver but provides reasonable ergonomics to handle client
changes at runtime.

We don't want to wipe state if the `Cluster` changes at all because
that will inevitably result in unnecessarily rebuilding state, which
can be an expensive operation. Furthermore, the spec also (by nature)
isn't adequate - it is primarily a context/namespace tuple, with the
actual config being loaded as part of boostrapping the K8s client.

The way this is now broken down:
 * Cluster reconciler internally uses `ConnectionManager` to track
   its state including clients + their connection timestamps
 * Consumers use `ClientManager` which allows getting a client for
   a given `Cluster` object and handles "versioning" in a reconciler
   friendly way
    * Internally, this defers to a `ClientProvider` interface to
      handle refreshing on-demand (i.e. compare tracked timestamp for
      `Cluster` object to canonical timestamp); the runtime impl for
      this is `ConnectionManager` while a `FakeClientProvider` is
      offered for tests so they don't need to run the actual
      `Cluster` reconciler