controllers: make the Cmd controller level-triggered rather than edge triggered
Before this PR, the controller would listen for process
changes and update the apiserver on each change.

After this PR, the controller does all bookkeeping about the state of the
process in local memory, then updates the apiserver only during reconciliation.

This should be way more robust and eliminate a lot of race conditions.