api: make ClientManager resource aware (#5352)
There's numerous competing concerns with managing K8s clients/their
operations, and a big one is that for efficiency, we only ever want
to set up a namespace-level watch for a specific cluster once and
then triage the incoming data.

In particular, this worked well with the engine subscribers, as
they'd always evaluate global state.

However, for reconcilers, this is trickier, as we operate on
individual objects. Behind the scenes, the reconciler tracks the
consumers for a given watch, so that it can de-duplicate and perform
refcounting to know when it's safe to cancel.

This complicates client refreshes, as if the underlying K8s client
for the `Cluster` object gets changed, we need to reset state for
each referencing object individually when its reconcile happens.

The changes here actually reduce the `ClientManager`'s surface area
by having it act purely as a coordinator on top of the
`ClientProvider` rather than an extra level of caching (i.e. it
no longer stores pointers to the clients themselves but always
delegates reads back to the provider).

As this coordination is now on a per-object basis, there's a slight
hack in the engine subscribers to use a single, virtual object for
keying so that it can continue acting on a global basis.