api: add retry + cleanup to PodSource
Currently, if the `PodSource` fails to watch a namespace, an error
will be logged, but there's no way to get out of that state because
the watch remains, so subsequent attempts to watch will short circuit.

Now, the watch logic uses a retry loop with backoff. Errors are logged
but still not propagated to `PodLogStream` objects currently. In
practice, this should be mostly fine because if the K8s client is
inoperable, the reconciler will also fail when it tries to actually
set up the log streaming. In the future, it'd probably be more robust
to track the error state and propagate it via `handleReconcileRequest`
so that the reconciler can mark the `PodLogStream` status accordingly
and then requeue itself.

Additionally, if there's no more watchers for a namespace, the watch
will be canceled. This will become more important with multi-cluster
work to ensure we dispose of resources for old clients.