api: report Docker connect events on Tiltfile load
After loading a Tiltfile, report the status of the Docker connection
if the Tiltfile iif it has Docker resources.

Currently, the Docker client is constructed very early in Tilt start
via DI (wire), and if an error is encountered, an exploding client is
used that will error out on any operation. It's not possible to
re-initialize the client at runtime (you need to restart Tilt), so the
status can never change. As a result, a `sync.Once` guards the metric
reporting so that it doesn't happen on every Tiltfile load.