cli: fix `tilt docker-prune` command panic
Re-use the logic from the reconciler directly to determine the
registry (based on the cluster + `default_registry` in Tiltfile)
in the CLI code.

This addresses a crash resulting from the image properties not being
present, as they're inferred + populated during Tiltfile
reconciliation, which does not occur here.

In the future, we'll likely want this (and other similar commands
such as `tilt alpha tiltfile-result`) to launch a headless apiserver
and wait for a condition, i.e. Tiltfile loaded, before reading the
objects from the apiserver to do work.