demo: fix perpetual "Waiting for cluster" in some setups (#5425)
This should always use the _local_ Docker client rather than the
cluster client because we're creating our own ephemeral cluster
and doing that e.g. inside of minikube's Docker is unlikely to
be what anyone wants beyond resulting in the UI getting perpetually
stuck at "Waiting for cluster" because it's launched an inaccessible
ephemeral cluster with comical levels of nested virtualization/
containerization.