engine: fix build graph structure error on LU-only targets
Historically, the engine builds a lightweight DAG for the targets
in a manifest. A sanity check is done up-front to ensure there's
a single root node (the "deploy" target, i.e. `K8sTarget` or
`DockerComposeTarget` for the cases relevant to this bug). That
node can then have children, which will be `ImageTarget`s, and
for completeness sake, those can recursively have further child
`ImageTarget`s for base images.

Live Update-only targets are currently modeled in the engine as
`ImageTarget`s with a special `IsLiveUpdateOnly` flag which causes
the engine to skip the (non-existant) build. Because there is no
build, there is no `ImageMap`.

As part of API migration, the deploy targets mentioned above are
now driven by their API specs as the source of truth, and both
`KubernetesApplySpec` and `DockerComposeUpSpec` include an
`ImageMaps` field, which is used when building the DAG.

This means that LU-only targets were no longer represented as a
dependency of the deploy target. In reality, this is correct!
However, the DAG sanity check now fails as a result. The change
here is to temporarily remove the LU-only targets from the set
before building the DAG, and then add them back in to the list
of deployed images.

Once API migration work is further along (e.g. Docker Compose
Live Update happens via reconciler), the DAG logic and this hack
will go away.