k8s: improve error messages during Pod exec failures
When running a command in a Pod, `client-go` attempts to handle
the failure case of a non-zero exit code by inspecting a
`metav1.Status` object.

For other failure cases, it returns `fmt.Errorf(status.Message)`,
and `Message` might be blank, resulting in cryptic error messages
from Tilt.

There's not a lot we can do without upstream `client-go` changes
on the Tilt side, so some extra context is added around where the
exec failed, and for the stream error case, a generic fallback
error message is used if we get an empty error from `client-go`.