cluster: refactor client initialization
The motivation here is in preparation for cluster status polling
to make it possible to write more robust unit tests. Currently,
the reconciler behaved very differently when provided with fake
clients for test. (In fact, there was actually a bug as a result
of this! The `ConnectedAt` timestamp handling was different for
tests.)

Now, there's factories for both K8s/Docker clients that get
injected, and tests can override these. Helper methods are
offered to return a static client for all created connections
or an error.

This latter part is critical to simulating errors with fake clients
in a controlled fashion to verify the (forthcoming) retry behavior.