tiltfile: prevent hang on extension fetch (#5610)
Extensions are fetched via Git, and it's possible for SSH to
decide to prompt the user (e.g. host key verification, key file
passphrase).

We don't have an interactive session, so this results in an
indefinite hang at startup while evalutating the Tiltfile.

Fix is to pull in tilt-dev/go-get#12, which sets `-o BatchMode yes`
on SSH invocations, so that it'll fail if it would result in a
prompt, which will be propagated back to the Tiltfile logs so the
user can resolve it.