cluster: monitor K8s livez/readyz after initialization (#5623)
After initializing a connection, start a monitor goroutine to
poll the `/livez` and `/readyz` endpoints of the cluster and
update the Tilt `Cluster` object status accordingly.

Additionally, if the client cannot be initialized in the first
place, the reconciler will requeue itself in the future and
attempt re-initialization.

Because many parts of Tilt still use a "fixed" cluster client
(i.e. dependency-injected Kubernetes client), some
functionality is gated behind a feature flag. Specifically,
if the cluster has an init error, no attempt at re-initing it
will be made unless the `cluster_refresh` flag is true. This
will allow us to continue migrating parts of the system to
use a dynamically obtained client based on the `Cluster`
object without impacting normal workflows.