script: use Go 1.18 image for Toast tasks
I changed `Dockerfile.toast` but forgot to update the reference
to the new image in `toast.yml`, so it was still using a 1.17
based image.

This meant that `make wire` would fail as it couldn't handle parsing
generics.

All tasks have been re-run as part of this commit to ensure they
are working and commit any minor changes resulting from the upgrade.