engine: remove Refs from ImageTarget
This is another place where we need to use the `Cluster` object
to determine the registry to resolve image references on the fly.

The brunt of the fallout here is actually throughout tests; many
of these now rely on the `.ImageMapSpec.Selector` directly, but
some that are testing specific registry-rewriting functionality
have been adapted to work.