cluster: log default registry information
This was previously in the Tiltfile reconciler as part of the
call to `DecideRegistry`, which is no longer used (and imminently
being removed).

Because it's logged on connection initialization, it will only
occur 1x per cluster client initialization. (That does mean that
if you change the spec and/or delete+recreate, it'll get logged
again once it re-initializes the client, but that's arguably a
good thing IMO!) The important part is that it doesn't log on
every reconcile, which could get spammy very fast.