Build dependency graph for k8s deletion in tilt down (#5668)
* Build dependency graph for k8s deletion

When running tilt down, the order of deletion isn't dependent on the
resource_deps specified. These changes introduce a dependency graph to
the down command to determine the correct deletion order.

* Refactors from code review

Make a function `sortManifestsForDeletion` to sort the manifests into
deletion order.

Add a map to the sort function to allow simpler lookup of dependency by
name.