cluster: improve unrecoverable error message
Until the `cluster_refresh` feature is done, initialization errors
for the cluster are unrecoverable, as not all parts of Tilt know
how to re-initialize the client/themsleves properly on change.

The error message is now wrapped with some additional context and
a notice that Tilt needs to be restarted after fixing the issue.