k8s_custom_deploy: fix Live Update rebuild
The `k8s_custom_deploy` is unusual: it's possible to use with
no Tilt-managed image build while still getting Live Update
functionality. This is useful for scenarios where you have an
all-in-one build&deploy tool (e.g. Knative) you want to use
with Tilt.

However, this meant that the value of the `deps` arg was being
passed both to the deploy target (K8s/`apply_cmd`) as well as
the image target (Live Update), which introduced a race between
them. If the Live Update reconciler saw a (supported) file change
first, it would clear it from the pending file list on the deploy
target once done, and so the deploy target would be skipped, as
it no longer had any pending changs. On the other hand, if the
deploy target saw it first, it'd start running because deploy
target changes can't be Live Update'd. At the same time, the Live
Update reconciler would still run, as it does not have a concept
of blocking/waiting for other targets.

To address this, the `deps` arg is now filtered using the generated
`LiveUpdateSpec` and any paths matching it (from `sync()` and
`fall_back_on` rules) are NOT included on the deploy target.
Similarly, the `FileWatch` for the `LiveUpdate` won't include paths
that it can't operate on.