ci: use Compose v2.4.1 for integration tests
We were pinned to v2.2.2 previously because v2.3.x had a regression
that made it unsuitable for use with Tilt, but v2.4.x is now out
(and shipping with latest Docker Desktop releases).

A couple of other bits were behind so I bumped them as well. (Note
that the Docker CLI is still behind because the latest static
binaries are mysteriously missing from the download site, but that's
not really a big deal.)