docker: fix spurious double builds on Dockerfile change
Do not _watch_ the Dockerfile from the image target, as this is
handled by the Tiltfile, which provides the full contents to the
target in-memory (i.e. the on-disk Dockerfile isn't actually used
when we execute the build).

This prevents race conditions from them both watching it, and
ensures that the flow always goes:
    (Dockerfile change) -> Tiltfile execute -> image build

Fixes #5586.