custom_build: add EXPECTED_{IMAGE,TAG,REGISTRY} env vars
Currently, we pass `EXPECTED_REF`, which is the full image reference
including registry, path, and tag.

Some build scripts want to consume these independently, so in addition
to passing the full reference, this adds the individual components as
well.

There's a slight behavior change here in that previously, the registry
was passed as `REGISTRY_HOST` only in the case that
`outputs_image_ref_to` mode was being used. Now, it's _always_ passed
(if there is a local registry!) regardless of custom build mode along
with `EXPECTED_REGISTRY`, which is new and duplicative but consistent
with the other names. This should be harmless, however.

Closes #5433.