registry: use API types everywhere
Eliminate the `container.Registry` helper and rely on the API type
of `v1alpha1.RegistryHosting`, which 1:1 maps to the local registry
spec.

A couple helpers methods are now functions, but the biggest change
is probably around default value semantics. For more consistency with
how the API types are actually used, the registry is `nil` rather
than empty.

The naming is also a bit more explicit for the changed areas, so that
"cluster ref" is now called "container runtime ref" to align with the
spec and avoid confusion between "cluster network ref", which is not
used by Tilt (aside from passing it along as metadata in the `Cluster`
API). I left a couple TODOs here for additional mechanical renames we
can do here but omitted them from this commit to keep it from getting
too unwieldy.