cluster: refactor health monitoring to avoid getting stuck (#5781)
I'm not 100% sure this is the root cause of the cluster health
monitoring getting stuck, but it's definitely not totally correct
in its current state.

The `ConnectionManager` is a fancy wrapper over a `sync.Map` and
expects the `connection` objects to be immutable — we replace them
entirely on change. Within the context of the reconciliation loop,
this is good. However, the health status monitor also needs to store
the result of the health checks and was doing so on the same object,
which resulted in a race condition.

Now, the cluster health state is stored independently and used within
reconciliation. The main `Reconcile()` loop still replaces objects
in their entirety in the `ConnectionManager`, but as its the only writer
now, there's no potential for stale data/races.

See #5729.