web: fix NodeJS engine version range
This was semantically invalid, the `,` should have been `||` to
support both 14.x and 16.x (the current LTS releases).

However, we only test/support building with v16.x, and there's no
compelling reason to support 14.x, as the Tilt UI is not intended
for usage as a library by other projects, and NodeJS is not a
_runtime_ dependency, only needed for build, so does not affect
end users (only those contributing to Tilt).

Closes #5317.