api: add conditions to KubernetesApply
Currently a single condition type is defined, which will be used
when a Job is applied that has already run successfully on the
cluster (and so is a no-op). This information will be propagated
so that Tilt can accurately determine the resource status despite
the potential non-existance of associated Pods.