store: update how we model file changes (#6503)
before this PR, we kept a map of recent file changes,
and deleted from the map when the changes were consumed.

this creates race conditions if the events are processed
in a different order than the changes, e.g.,
if a "consumed" event at 12:02 is processed before
the file change event at 12:01.

after this PR, we record both the file change time
and the consumed time, then compare them to determine
if the file change is still pending.

Signed-off-by: Nick Santos <nick.santos@docker.com>