test: more fs mocks (#6511)
using a real fs is causing some test flakiness, because sometimes controllers
keep writing to the filesystem after the test ends.

another way to fix this would be to block
the clean up of every test on reconciler
shutdown. that's probably a good thing to
do too eventually but am not sure how much
it will slow things down.

Signed-off-by: Nick Santos <nick.santos@docker.com>