When calculating Kustomize dependencies, include components (#6534)
When calling "kustomize(<directory>)" in a Tiltfile, the directory
is watched as well as any paths the Kustomization includes. This
doesn't currently include components, so Kustomizations that
include components[1] don't trigger a reload of the Tiltfile
and the Kustomization when files in those components change.

This commit adds the "components:"" field of a Kustomization to
the list of paths "dependenciesForKustomization" will inspect.

[1]: https://github.com/kubernetes-sigs/kustomize/blob/master/examples/components.md

Signed-off-by: Sam Clements <sam@borntyping.co.uk>